utils: extract OrderEnum logic
